Change price and priceBeforeSale type to BigDecimal

+++ src/main/java/models/ProductModel.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -13,9 +15,9 @@ import lombok.Setter;
 public class ProductModel {
 	public final String name;
 	public final Currency priceCurrency;
-	public final Double price;
+	public final BigDecimal price;
 	public final Boolean isOnSale;
-	public final Double priceBeforeSale;
+	public final BigDecimal priceBeforeSale;
 	public final Boolean isOutOfStock;
 	public final String imageUrl;
 }
